refactor: split media transcription and callback domain service
