+++ backend-spring/src/main/java/com/myway/backendspring/feature/FeatureStoreService.java
@@ -5,6 +5,7 @@ import com.myway.backendspring.domain.LectureItem;
 import com.myway.backendspring.domain.ActivityEventService;
 import com.myway.backendspring.feature.ai.AiFeatureService;
 import com.myway.backendspring.feature.media.MediaProcessingService;
+import com.myway.backendspring.feature.media.MediaTranscriptionService;
 import com.myway.backendspring.feature.quota.AiUsageQuotaService;
 import com.myway.backendspring.feature.repository.FeatureStoreRepository;
 import com.myway.backendspring.feature.rag.RagService;
@@ -42,14 +43,8 @@ public class FeatureStoreService {
     private static final String ADMIN_ASSIGNMENT_SCOPE = "admin_assignment";
     private static final String AI_USAGE_SCOPE = "ai_usage_daily";
     private static final String RAG_INDEX_SCOPE = "rag_chunk_index";
-    private static final String STT_DEFAULT_PROVIDER = "cloudflare";
-    private static final String STT_DEFAULT_MODEL = "cf-whisper";
-    private static final int PUBLIC_STT_MAX_DURATION_MS = 180_000;
-    private static final int FALLBACK_TRANSCRIPT_DURATION_MS = 120_000;
-    private static final int FALLBACK_SEGMENT_MIN_MS = 1_000;
     private static final int RAG_CHUNK_MAX_WORDS = 90;
     private static final int RAG_CHUNK_OVERLAP_WORDS = 20;
-    private static final int STT_TARGET_SEGMENT_WORDS = 20;
     private final FeatureJdbcStore store;
     private final DemoLearningService learningService;
     private final int shortformMaxRetry;
@@ -63,6 +58,7 @@ public class FeatureStoreService {
     private final ShortformService shortformService;
     private final AiFeatureService aiFeatureService;
     private final MediaProcessingService mediaProcessingService;
+    private final MediaTranscriptionService mediaTranscriptionService;
     private final ObjectMapper objectMapper = new ObjectMapper();
 
     @Autowired
@@ -79,7 +75,8 @@ public class FeatureStoreService {
             AiUsageQuotaService aiUsageQuotaService,
             ShortformService shortformService,
             AiFeatureService aiFeatureService,
-            MediaProcessingService mediaProcessingService
+            MediaProcessingService mediaProcessingService,
+            MediaTranscriptionService mediaTranscriptionService
     ) {
         this.store = store;
         this.learningService = learningService;
@@ -94,6 +91,7 @@ public class FeatureStoreService {
         this.shortformService = shortformService;
         this.aiFeatureService = aiFeatureService;
         this.mediaProcessingService = mediaProcessingService;
+        this.mediaTranscriptionService = mediaTranscriptionService;
     }
 
     // Backward-compatible constructor for tests instantiating service directly.
@@ -111,6 +109,7 @@ public class FeatureStoreService {
                 null,
                 null,
                 new AiFeatureService(new FeatureStoreRepository(store), null, "dev"),
+                null,
                 null
         );
     }
@@ -284,102 +283,14 @@ public class FeatureStoreService {
     }
 
     public Map<String, Object> transcribe(String lectureId, String language, Integer durationMsInput, String sttProvider, String sttModel, String audioUrl, String extractionId) {
+        if (mediaTranscriptionService == null) return Map.of();
         Map<String, Object> extraction = extractionId == null || extractionId.isBlank()
                 ? createExtraction(lectureId, audioUrl)
                 : store.getKv(EXTRACTION_SCOPE, extractionId);
         if (extraction == null) {
             extraction = createExtraction(lectureId, audioUrl);
         }
-        String now = Instant.now().toString();
-        Map<String, Object> extractionInProgress = new HashMap<>(extraction);
-        extractionInProgress.put("status", "PROCESSING");
-        extractionInProgress.put("stt_status", "PROCESSING");
-        extractionInProgress.put("processing_stage", "transcribing");
-        extractionInProgress.put("processing_step", "stt_started");
-        extractionInProgress.put("processing_error_code", null);
-        extractionInProgress.put("processing_error", null);
-        extractionInProgress.put("updated_at", now);
-        store.upsertKv(EXTRACTION_SCOPE, String.valueOf(extractionInProgress.getOrDefault("id", extractionId)), extractionInProgress);
-        extraction = extractionInProgress;
-        LectureItem lecture = learningService.getLecture(lectureId);
-        int lectureDurationMs = lecture == null
-                ? FALLBACK_TRANSCRIPT_DURATION_MS
-                : Math.max(FALLBACK_SEGMENT_MIN_MS * 3, lecture.duration_minutes() * 60_000);
-        int requestedDurationMs = durationMsInput == null || durationMsInput <= 0 ? lectureDurationMs : durationMsInput;
-        int durationMs = Math.min(requestedDurationMs, PUBLIC_STT_MAX_DURATION_MS);
-
-        String normalizedLanguage = language == null || language.isBlank() ? "ko" : language;
-        String resolvedProvider = sttProvider == null || sttProvider.isBlank() ? STT_DEFAULT_PROVIDER : sttProvider.trim();
-        String resolvedModel = sttModel == null || sttModel.isBlank() ? STT_DEFAULT_MODEL : sttModel.trim();
-        String baseText = buildLectureNarrative(lectureId);
-        List<Map<String, Object>> segments = splitIntoSegments(baseText, durationMs);
-        String transcriptId = String.valueOf(extraction.getOrDefault("id", UUID.randomUUID().toString()));
-        int wordCount = countWords(baseText);
-        Map<String, Object> sttQuality = sttQualityMetrics(segments, durationMs);
-
-        Map<String, Object> transcriptPayload = new HashMap<>();
-        transcriptPayload.put("id", transcriptId);
-        transcriptPayload.put("lecture_id", lectureId);
-        transcriptPayload.put("language", normalizedLanguage);
-        transcriptPayload.put("full_text", baseText);
-        transcriptPayload.put("segments", segments);
-        transcriptPayload.put("word_count", wordCount);
-        transcriptPayload.put("duration_ms", durationMs);
-        transcriptPayload.put("stt_provider", resolvedProvider);
-        transcriptPayload.put("stt_model", resolvedModel);
-        transcriptPayload.put("quality", sttQuality);
-        transcriptPayload.put("created_at", Instant.now().toString());
-        store.upsertKv(TRANSCRIPT_SCOPE, lectureId, transcriptPayload);
-
-        Map<String, Object> pipelinePayload = new HashMap<>();
-        pipelinePayload.put("lecture_id", lectureId);
-        pipelinePayload.put("transcript_status", "COMPLETED");
-        pipelinePayload.put("summary_status", "PENDING");
-        pipelinePayload.put("audio_status", "COMPLETED");
-        pipelinePayload.put("processing_stage", "completed");
-        pipelinePayload.put("processing_step", "stt_completed");
-        pipelinePayload.put("processing_error_code", null);
-        pipelinePayload.put("processing_error", null);
-        pipelinePayload.put("transcript_id", transcriptId);
-        pipelinePayload.put("note_id", null);
-        pipelinePayload.put("extraction_id", extraction.get("id"));
-        pipelinePayload.put("updated_at", now);
-        store.upsertKv(PIPELINE_SCOPE, lectureId, pipelinePayload);
-
-        Map<String, Object> extractionPayload = new HashMap<>(extraction);
-        extractionPayload.put("status", "COMPLETED");
-        extractionPayload.put("stt_status", "COMPLETED");
-        extractionPayload.put("transcript_id", transcriptId);
-        extractionPayload.put("audio_duration_ms", durationMs);
-        extractionPayload.put("processed_at", now);
-        extractionPayload.put("updated_at", now);
-        extractionPayload.put("language", normalizedLanguage);
-        extractionPayload.put("requested_stt_provider", resolvedProvider);
-        extractionPayload.put("requested_stt_model", resolvedModel);
-        extractionPayload.put("audio_url", audioUrl);
-        extractionPayload.put("processing_stage", "completed");
-        extractionPayload.put("processing_step", "stt_completed");
-        extractionPayload.put("processing_error_code", null);
-        extractionPayload.put("processing_error", null);
-        extractionPayload.put("stt_quality", sttQuality);
-        store.upsertKv(EXTRACTION_SCOPE, String.valueOf(extraction.getOrDefault("id", transcriptId)), extractionPayload);
-
-        Map<String, Object> ragIndex = rebuildRagIndex(lectureId, null);
-
-        Map<String, Object> response = new HashMap<>();
-        response.put("transcript_id", transcriptId);
-        response.put("lecture_id", lectureId);
-        response.put("segment_count", segments.size());
-        response.put("duration_ms", durationMs);
-        response.put("word_count", wordCount);
-        response.put("stt_provider", resolvedProvider);
-        response.put("stt_model", resolvedModel);
-        response.put("audio_url", audioUrl);
-        response.put("pipeline", pipeline(lectureId));
-        response.put("language", normalizedLanguage);
-        response.put("quality", sttQuality);
-        response.put("rag_index", ragIndex);
-        return response;
+        return mediaTranscriptionService.transcribe(extraction, lectureId, language, durationMsInput, sttProvider, sttModel, audioUrl);
     }
 
     public List<Map<String, Object>> extractions(String lectureId) {
@@ -420,127 +331,20 @@ public class FeatureStoreService {
             Integer sampleRate,
             Integer channels
     ) {
-        Map<String, Object> extraction = store.getKv(EXTRACTION_SCOPE, extractionId);
-        if (extraction != null) {
-            extraction = new HashMap<>(extraction);
-        }
-        if (extraction == null) {
-            for (Map<String, Object> row : store.listEventsByScope(EXTRACTION_SCOPE)) {
-                if (extractionId.equals(String.valueOf(row.getOrDefault("id", "")))) {
-                    extraction = new HashMap<>(row);
-                    break;
-                }
-            }
-        }
-        if (extraction == null) {
-            return null;
-        }
-
-        long currentVersion = asLong(extraction.get("last_event_version"));
-        if (eventVersion <= currentVersion) {
-            extraction.put("callback_ignored", true);
-            extraction.put("callback_status", "IGNORED_STALE");
-            return normalizeExtractionForResponse(extraction);
-        }
-
-        String now = Instant.now().toString();
-        String resolvedStatus = status == null || status.isBlank() ? "COMPLETED" : status.toUpperCase();
-        String resolvedErrorCode = "FAILED".equalsIgnoreCase(resolvedStatus) ? "PROCESSOR_CALLBACK_FAILED" : null;
-        if ("FAILED".equalsIgnoreCase(resolvedStatus) && (errorMessage == null || errorMessage.isBlank())) {
-            errorMessage = "media processor callback failed";
-        }
-        extraction.put("last_event_version", eventVersion);
-        extraction.put("status", resolvedStatus);
-        extraction.put("error_message", errorMessage);
-        if (audioUrl != null && !audioUrl.isBlank()) {
-            extraction.put("audio_url", audioUrl);
-        }
-        if (processingJobId != null && !processingJobId.isBlank()) {
-            extraction.put("processing_job_id", processingJobId);
-        }
-        if (processingStage != null && !processingStage.isBlank()) {
-            extraction.put("processing_stage", processingStage);
-        }
-        if (processingStep != null && !processingStep.isBlank()) {
-            extraction.put("processing_step", processingStep);
-        }
-        if (audioFormat != null && !audioFormat.isBlank()) {
-            extraction.put("audio_format", audioFormat);
-        }
-        if (sampleRate != null && sampleRate > 0) {
-            extraction.put("sample_rate", sampleRate);
-        }
-        if (channels != null && channels > 0) {
-            extraction.put("channels", channels);
-        }
-        if ("COMPLETED".equalsIgnoreCase(resolvedStatus)) {
-            extraction.put("status", "PROCESSING");
-            extraction.put("processing_stage", "transcribing");
-            extraction.put("processing_step", "stt_started");
-            extraction.put("stt_status", "PROCESSING");
-            extraction.put("processing_error_code", null);
-            extraction.put("processing_error", null);
-            extraction.put("processed_at", now);
-        } else if ("FAILED".equalsIgnoreCase(resolvedStatus)) {
-            extraction.put("processing_stage", "failed");
-            extraction.put("processing_step", "callback_failed");
-            extraction.put("stt_status", "FAILED");
-            extraction.put("processing_error_code", resolvedErrorCode);
-            extraction.put("processing_error", errorMessage);
-            extraction.put("processed_at", now);
-        } else {
-            extraction.put("processing_stage", "callback");
-            extraction.put("processing_step", "callback_received");
-        }
-        extraction.put("callback_ignored", false);
-        extraction.put("callback_status", "APPLIED");
-        extraction.put("updated_at", now);
-        store.upsertKv(EXTRACTION_SCOPE, extractionId, extraction);
-
-        String lectureId = String.valueOf(extraction.getOrDefault("lecture_id", "")).trim();
-        if (!lectureId.isBlank()) {
-            Map<String, Object> pipeline = new HashMap<>();
-            pipeline.put("lecture_id", lectureId);
-            pipeline.put("audio_status", "FAILED".equalsIgnoreCase(resolvedStatus) ? "FAILED" : "COMPLETED");
-            pipeline.put("transcript_status", "FAILED".equalsIgnoreCase(resolvedStatus) ? "FAILED" : "PROCESSING");
-            pipeline.put("summary_status", "PENDING");
-            pipeline.put("processing_stage", "FAILED".equalsIgnoreCase(resolvedStatus) ? "failed" : "transcribing");
-            pipeline.put("processing_step", "FAILED".equalsIgnoreCase(resolvedStatus) ? "callback_failed" : "stt_started");
-            pipeline.put("processing_error_code", "FAILED".equalsIgnoreCase(resolvedStatus) ? resolvedErrorCode : null);
-            pipeline.put("processing_error", "FAILED".equalsIgnoreCase(resolvedStatus) ? errorMessage : null);
-            pipeline.put("transcript_id", extraction.get("transcript_id"));
-            pipeline.put("note_id", null);
-            pipeline.put("extraction_id", extraction.get("id"));
-            pipeline.put("updated_at", now);
-            store.upsertKv(PIPELINE_SCOPE, lectureId, pipeline);
-        }
-        if (activityEventService != null) {
-            String userId = String.valueOf(extraction.getOrDefault("user_id", "")).trim();
-            if (!userId.isBlank()) {
-                activityEventService.append(
-                        userId,
-                        "media_extraction_" + resolvedStatus.toLowerCase(),
-                        "extraction",
-                        extractionId,
-                        Map.of("lecture_id", lectureId, "status", resolvedStatus)
-                );
-            }
-        }
-        return normalizeExtractionForResponse(extraction);
-    }
-
-    private Map<String, Object> normalizeExtractionForResponse(Map<String, Object> extraction) {
-        Map<String, Object> hydrated = new HashMap<>(extraction);
-        hydrated.putIfAbsent("processing_stage", "queued");
-        hydrated.putIfAbsent("processing_step", "job_requested");
-        hydrated.putIfAbsent("processing_error_code", null);
-        hydrated.putIfAbsent("processing_error", null);
-        hydrated.putIfAbsent("stt_status", "PENDING");
-        hydrated.putIfAbsent("error_message", null);
-        hydrated.putIfAbsent("callback_ignored", false);
-        hydrated.putIfAbsent("callback_status", "APPLIED");
-        hydrated.putIfAbsent("last_event_version", 0L);
-        return hydrated;
+        if (mediaTranscriptionService == null) return null;
+        return mediaTranscriptionService.completeExtractionCallback(
+                extractionId,
+                status,
+                errorMessage,
+                eventVersion,
+                audioUrl,
+                processingJobId,
+                processingStage,
+                processingStep,
+                audioFormat,
+                sampleRate,
+                channels
+        );
     }
 
     public Map<String, Object> pipeline(String lectureId) {
@@ -724,20 +528,6 @@ public class FeatureStoreService {
         return payload;
     }
 
-    private int asInt(Object value) {
-        if (value == null) {
-            return 0;
-        }
-        if (value instanceof Number n) {
-            return n.intValue();
-        }
-        try {
-            return Integer.parseInt(String.valueOf(value));
-        } catch (Exception ignored) {
-            return 0;
-        }
-    }
-
     private boolean asBoolean(Object value, boolean defaultValue) {
         if (value == null) {
             return defaultValue;
@@ -756,20 +546,6 @@ public class FeatureStoreService {
             return false;
         }
         return defaultValue;
-    }
-
-    private long asLong(Object value) {
-        if (value == null) {
-            return 0L;
-        }
-        if (value instanceof Number n) {
-            return n.longValue();
-        }
-        try {
-            return Long.parseLong(String.valueOf(value));
-        } catch (Exception ignored) {
-            return 0L;
-        }
     }
 
     private String usageKey(String userId) {
@@ -985,71 +761,6 @@ public class FeatureStoreService {
         }
         return lecture.title() + " 강의에서는 핵심 개념을 설명하고, 실습 흐름과 복습 포인트를 단계적으로 정리합니다. " +
                 "특히 자주 혼동되는 개념을 예시로 비교하고, 다음 차시와 연결되는 질문을 중심으로 학습합니다.";
-    }
-
-    private List<Map<String, Object>> splitIntoSegments(String text, int durationMs) {
-        List<String> words = tokenizeForChunking(text);
-        if (words.isEmpty()) {
-            return List.of();
-        }
-        int segmentCount = Math.min(8, Math.max(3, (int) Math.ceil(words.size() / 20.0)));
-        int wordsPerSegment = Math.max(1, (int) Math.ceil(words.size() / (double) segmentCount));
-        List<Map<String, Object>> segments = new ArrayList<>();
-        for (int i = 0; i < segmentCount; i++) {
-            int start = i * wordsPerSegment;
-            int end = Math.min(words.size(), start + wordsPerSegment);
-            if (start >= end) {
-                break;
-            }
-            int startMs = (int) Math.round((i / (double) segmentCount) * durationMs);
-            int endMs = i == segmentCount - 1
-                    ? durationMs
-                    : (int) Math.round(((i + 1) / (double) segmentCount) * durationMs);
-            String segmentText = String.join(" ", words.subList(start, end));
-            Map<String, Object> segment = new HashMap<>();
-            segment.put("index", i);
-            segment.put("start_ms", startMs);
-            segment.put("end_ms", Math.max(startMs + FALLBACK_SEGMENT_MIN_MS, endMs));
-            segment.put("text", segmentText);
-            segments.add(segment);
-        }
-        return segments;
-    }
-
-    private Map<String, Object> sttQualityMetrics(List<Map<String, Object>> segments, int durationMs) {
-        if (segments == null || segments.isEmpty()) {
-            return Map.of(
-                    "segment_count", 0,
-                    "avg_words_per_segment", 0,
-                    "min_segment_ms", 0,
-                    "max_segment_ms", 0,
-                    "target_segment_words", STT_TARGET_SEGMENT_WORDS,
-                    "quality_score", 0.0
-            );
-        }
-        int totalWords = 0;
-        int minMs = Integer.MAX_VALUE;
-        int maxMs = 0;
-        for (Map<String, Object> segment : segments) {
-            int start = asInt(segment.get("start_ms"));
-            int end = asInt(segment.get("end_ms"));
-            int ms = Math.max(0, end - start);
-            minMs = Math.min(minMs, ms);
-            maxMs = Math.max(maxMs, ms);
-            totalWords += countWords(String.valueOf(segment.getOrDefault("text", "")));
-        }
-        double avgWords = totalWords / (double) Math.max(1, segments.size());
-        double wordFit = 1.0 - Math.min(1.0, Math.abs(avgWords - STT_TARGET_SEGMENT_WORDS) / STT_TARGET_SEGMENT_WORDS);
-        double durationFit = 1.0 - Math.min(1.0, Math.abs(durationMs - (segments.size() * 20_000.0)) / Math.max(1.0, durationMs));
-        double score = Math.max(0.0, Math.min(1.0, (wordFit * 0.7) + (durationFit * 0.3)));
-        return Map.of(
-                "segment_count", segments.size(),
-                "avg_words_per_segment", Math.round(avgWords * 100.0) / 100.0,
-                "min_segment_ms", minMs == Integer.MAX_VALUE ? 0 : minMs,
-                "max_segment_ms", maxMs,
-                "target_segment_words", STT_TARGET_SEGMENT_WORDS,
-                "quality_score", Math.round(score * 1000.0) / 1000.0
-        );
     }
 
     private int countWords(String text) {
